fix: 산책일지 삭제, 강아지 생성/수정/삭제시에도 statistics 데이터 업데이트하도록 변경
- 위의 경우에 업데이트를 하지 않아 변경사항이 홈 화면에 반영되지 않는 문제가 있었다
- const 폴더를 생성하고 CACHE 관련 const를 cache-const.ts 파일에 이동시켰다. 다른 const들도 이 폴
더에 모으려고 함

+++ backend/server/src/journals/journals.service.ts
@@ -7,24 +7,25 @@ import { Transactional } from 'typeorm-transactional';
 import { Journals } from './journals.entity';
 import { JournalsRepository } from './journals.repository';
 import {
-    ExcrementsInputForCreate,
+    CreateJournalDatabaseInput,
     CreateJournalRequest,
     DogOutputForDetail,
     DogWalkJournalRaw,
     ExcrementCount,
-    JournalDetailResponse,
+    ExcrementsInputForCreate,
     JournalDetailRaw,
-    JournalOutputForDetail,
+    JournalDetailResponse,
+    JournalInputForCreate,
     JournalListResponse,
+    JournalOutputForDetail,
+    UpdateJournalDatabaseInput,
     UpdateJournalRequest,
     UpdateTodayWalkTimeOperation,
-    UpdateJournalDatabaseInput,
-    CreateJournalDatabaseInput,
-    JournalInputForCreate,
 } from './types/journal.types';
 
 import { WinstonLoggerService } from '../common/logger/winstonLogger.service';
 
+import { EVENTS } from '../const/cache-const';
 import { DogWalkDayService } from '../dog-walk-day/dog-walk-day.service';
 import { DogsService } from '../dogs/dogs.service';
 import { Excrements } from '../excrements/excrements.entity';
@@ -202,7 +203,7 @@ export class JournalsService {
             await this.createExcrements(createJournalResult.id, createJournalRequest.excrements);
         }
 
-        this.eventEmitter.emit('journal.created', { userId });
+        await this.eventEmitter.emit(EVENTS.JOURNAL_CREATED, { userId });
     }
 
     @Transactional()
@@ -228,8 +229,12 @@ export class JournalsService {
         await this.updateDogWalkDay(dogIds, subtractDogWalkDay);
         await this.updateTodayWalkTime(dogIds, journalRaw.duration, subtractTodayWalkTime);
 
-        await this.s3Service.deleteObjects(userId, journalPhotos);
+        if (journalPhotos.length) {
+            await this.s3Service.deleteObjects(userId, journalPhotos);
+        }
         await this.delete(journalId);
+
+        await this.eventEmitter.emit(EVENTS.JOURNAL_DELETED, { userId });
     }
 
     private async findUserDogJournalsByDate(
